Finished (hopefully) the URI comparison
