+++ src/FubarDev.WebDavServer/Utils/DefaultUriComparer.cs
@@ -5,8 +5,9 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+
+using FubarDev.WebDavServer.Properties;
 
 namespace FubarDev.WebDavServer.Utils
 {
@@ -15,6 +16,7 @@ namespace FubarDev.WebDavServer.Utils
     /// </summary>
     public class DefaultUriComparer : IUriComparer
     {
+        private static readonly char[] _slash = { '/' };
         private readonly IWebDavContextAccessor? _contextAccessor;
 
         /// <summary>
@@ -29,127 +31,111 @@ namespace FubarDev.WebDavServer.Utils
         /// <inheritdoc />
         public UriComparisonResult Compare(Uri x, Uri y)
         {
-            if (TryGetAbsolute(x, out var absoluteX))
-            {
-                x = absoluteX;
-            }
-
-            if (TryGetAbsolute(y, out var absoluteY))
-            {
-                y = absoluteY;
-            }
-
-            if (!IsSameHost(x, y))
-            {
-                return UriComparisonResult.PrecedingDifferentHost;
-            }
-
-            var context = _contextAccessor?.WebDavContext;
-            var uriX = MakeAbsoluteUri(x, context);
-            var uriY = MakeAbsoluteUri(y, context);
-
-            if (uriX is null)
-            {
-                if (uriY is null)
-                {
-                    // Both are relative URLs, so let's make them absolute
-                    return CompareAbsoluteUris(
-                        new Uri("/" + x.OriginalString),
-                        new Uri("/" + y.OriginalString));
-                }
-
-                return InvertResult(CompareWithAbsoluteUri(uriY, x));
-            }
-
-            if (uriY is null)
-            {
-                return CompareWithAbsoluteUri(uriX, y);
-            }
-
-            return CompareAbsoluteUris(uriX, uriY);
+            var publicBaseUrl = _contextAccessor?.WebDavContext.PublicBaseUrl;
+            var infoX = new UriInfo(x, publicBaseUrl);
+            var infoY = new UriInfo(y, publicBaseUrl);
+            return Compare(infoX, infoY);
         }
 
         /// <inheritdoc />
         public bool IsSameHost(Uri x, Uri y)
         {
-            if (!TryGetAbsolute(x, out var absoluteX)
-                || !TryGetAbsolute(y, out var absoluteY)
-                || string.IsNullOrEmpty(absoluteX.Host)
-                || string.IsNullOrEmpty(absoluteY.Host))
-            {
-                // It's the same host if one of them are relative URIs
-                // or the host is empty
-                return true;
-            }
-
-            x = absoluteX;
-            y = absoluteY;
-
-            var uriComponents = string.IsNullOrEmpty(x.Scheme) || string.IsNullOrEmpty(y.Scheme)
-                ? UriComponents.Host | UriComponents.Port
-                : UriComponents.Scheme | UriComponents.HostAndPort;
-            var valueX = x.GetComponents(uriComponents, UriFormat.UriEscaped);
-            var valueY = y.GetComponents(uriComponents, UriFormat.UriEscaped);
-
-            return string.Equals(valueX, valueY, StringComparison.OrdinalIgnoreCase);
+            var publicBaseUrl = _contextAccessor?.WebDavContext.PublicBaseUrl;
+            var infoX = new UriInfo(x, publicBaseUrl);
+            var infoY = new UriInfo(y, publicBaseUrl);
+            return CompareHosts(infoX, infoY) == UriComparisonResult.Equal;
         }
 
         /// <inheritdoc />
         public bool IsThisServer(Uri uri)
         {
-            return IsThisServer(uri, null);
+            return IsThisServer(uri, _contextAccessor?.WebDavContext.PublicBaseUrl);
         }
 
         /// <inheritdoc />
-        int IComparer<Uri>.Compare(Uri x, Uri y)
+        int IComparer<Uri>.Compare(Uri? x, Uri? y)
         {
+            if (x is null)
+            {
+                if (y is null)
+                {
+                    return 0;
+                }
+
+                return -10;
+            }
+
+            if (y is null)
+            {
+                return 10;
+            }
+
             return (int)Compare(x, y);
+        }
+
+        private static UriComparisonResult CompareRelativePaths(string pathX, string pathY)
+        {
+            var partsX = pathX.Split(_slash, StringSplitOptions.RemoveEmptyEntries);
+            var partsY = pathY.Split(_slash, StringSplitOptions.RemoveEmptyEntries);
+            var minLength = Math.Min(partsX.Length, partsY.Length);
+            for (var i = 0; i != minLength; ++i)
+            {
+                var partX = partsX[i];
+                var partY = partsY[i];
+                var partCompare = string.Compare(partX, partY, StringComparison.OrdinalIgnoreCase);
+                switch (partCompare)
+                {
+                    case < 0:
+                        return UriComparisonResult.FollowingSibling;
+                    case > 0:
+                        return UriComparisonResult.PrecedingSibling;
+                }
+            }
+
+            return partsX.Length == partsY.Length
+                ? UriComparisonResult.Equal
+                : partsX.Length > partsY.Length
+                    ? UriComparisonResult.Child
+                    : UriComparisonResult.Parent;
+        }
+
+        private static UriComparisonResult CompareHosts(UriInfo infoX, UriInfo infoY)
+        {
+            if (!infoX.IsAbsolute
+                || !infoY.IsAbsolute
+                || string.IsNullOrEmpty(infoX.Uri.Host)
+                || string.IsNullOrEmpty(infoY.Uri.Host))
+            {
+                // It's the same host if one of them are relative URIs
+                // or the host is empty
+                return UriComparisonResult.Equal;
+            }
+
+            var uriComponents = string.IsNullOrEmpty(infoX.Scheme) || string.IsNullOrEmpty(infoX.Scheme)
+                ? UriComponents.Host | UriComponents.Port
+                : UriComponents.Scheme | UriComponents.HostAndPort;
+            var valueX = infoX.Uri.GetComponents(uriComponents, UriFormat.UriEscaped);
+            var valueY = infoY.Uri.GetComponents(uriComponents, UriFormat.UriEscaped);
+
+            return string.Compare(valueX, valueY, StringComparison.OrdinalIgnoreCase) switch
+            {
+                < 0 => UriComparisonResult.PrecedingDifferentHost,
+                0 => UriComparisonResult.Equal,
+                > 0 => UriComparisonResult.FollowingDifferentHost,
+            };
         }
 
         private static UriComparisonResult InvertResult(UriComparisonResult result)
             => result switch
             {
+                UriComparisonResult.PrecedingDifferentHost => UriComparisonResult.FollowingDifferentHost,
+                UriComparisonResult.FollowingDifferentHost => UriComparisonResult.PrecedingDifferentHost,
                 UriComparisonResult.PrecedingSibling => UriComparisonResult.FollowingSibling,
                 UriComparisonResult.FollowingSibling => UriComparisonResult.PrecedingSibling,
                 UriComparisonResult.Parent => UriComparisonResult.Child,
                 UriComparisonResult.Child => UriComparisonResult.Parent,
                 _ => result,
             };
-
-        private static string FixUriPath(string path)
-        {
-            path = path.TrimEnd('/');
-            if (string.IsNullOrEmpty(path))
-            {
-                return "/";
-            }
-
-            if (!path.StartsWith("/"))
-            {
-                return "/" + path;
-            }
-
-            return path;
-        }
-
-        private static bool TryGetAbsolute(Uri uri, [NotNullWhen(true)] out Uri? absoluteUri)
-        {
-            if (uri.IsAbsoluteUri)
-            {
-                absoluteUri = uri;
-                return true;
-            }
-
-            var originalString = uri.OriginalString;
-            if (originalString.StartsWith("/"))
-            {
-                absoluteUri = new Uri(originalString);
-                return true;
-            }
-
-            absoluteUri = null;
-            return false;
-        }
 
         /// <summary>
         /// Compares a relative URL with an absolute URL.
@@ -162,18 +148,18 @@ namespace FubarDev.WebDavServer.Utils
         /// <remarks>
         /// This is all just guesswork!
         /// </remarks>
-        private UriComparisonResult CompareWithAbsoluteUri(
+        private static UriComparisonResult CompareWithAbsoluteUri(
             Uri absolute,
             Uri relative)
         {
             if (!absolute.IsAbsoluteUri)
             {
-                throw new ArgumentException("The URI must be absolute", nameof(absolute));
+                throw new ArgumentException(Resources.UriMustBeAbsolute, nameof(absolute));
             }
 
-            if (!relative.IsAbsoluteUri)
+            if (relative.IsAbsoluteUri)
             {
-                throw new ArgumentException("The URI must not be absolute", nameof(relative));
+                throw new ArgumentException(Resources.UriMustNotBeAbsolute, nameof(relative));
             }
 
             var relativePathParts = relative.OriginalString
@@ -254,35 +240,34 @@ namespace FubarDev.WebDavServer.Utils
                 : UriComparisonResult.FollowingSibling;
         }
 
-        private UriComparisonResult CompareAbsoluteUris(
-            Uri x,
-            Uri y)
+        private static UriComparisonResult Compare(UriInfo infoX, UriInfo infoY)
         {
-            if (!x.IsAbsoluteUri)
+            if (!infoX.IsAbsolute && !infoY.IsAbsolute)
             {
-                throw new ArgumentException("The URI must be absolute", nameof(x));
+                return CompareRelativePaths(infoX.PathWithSlash, infoY.PathWithSlash);
             }
 
-            if (!y.IsAbsoluteUri)
+            var hostCompareResult = CompareHosts(infoX, infoY);
+            if (hostCompareResult != UriComparisonResult.Equal)
             {
-                throw new ArgumentException("The URI must be absolute", nameof(y));
+                return hostCompareResult;
             }
 
-            var pathX = FixUriPath(x.GetComponents(UriComponents.Path, UriFormat.UriEscaped));
-            var pathY = FixUriPath(y.GetComponents(UriComponents.Path, UriFormat.UriEscaped));
-            if (string.Equals(pathX, pathY, StringComparison.OrdinalIgnoreCase))
+            if (!infoX.IsAbsolute)
             {
-                return UriComparisonResult.Equal;
+                return InvertResult(CompareWithAbsoluteUri(infoY.Uri, infoX.Uri));
             }
 
-            var uriX = new Uri(pathX);
-            var uriY = new Uri(pathY);
-            return uriX.IsBaseOf(uriY)
-                ? UriComparisonResult.Child
-                : UriComparisonResult.Parent;
+            if (!infoY.IsAbsolute)
+            {
+                return CompareWithAbsoluteUri(infoX.Uri, infoY.Uri);
+            }
+
+            // Just assume that the root path is the root path of the WebDAV service
+            return CompareRelativePaths(infoX.PathWithSlash, infoY.PathWithSlash);
         }
 
-        private bool IsThisServer(Uri uri, IWebDavContext? context)
+        private bool IsThisServer(Uri uri, Uri? publicBaseUrl)
         {
             if (_contextAccessor is null)
             {
@@ -290,81 +275,96 @@ namespace FubarDev.WebDavServer.Utils
                 return true;
             }
 
-            if (!TryGetAbsolute(uri, out var absoluteUri))
+            var uriInfo = new UriInfo(uri, publicBaseUrl);
+            if (!uriInfo.IsAbsolute || publicBaseUrl is null)
             {
-                // Relative URIs are always the same server
+                // URL is not absolute
                 return true;
             }
 
-            uri = absoluteUri;
-
-            context ??= _contextAccessor.WebDavContext;
-            var publicBaseUrl = context.PublicBaseUrl;
-
-            if (string.IsNullOrEmpty(uri.Host) && uri.Scheme == "file" && uri.AbsolutePath.StartsWith("//"))
-            {
-                // This is a relative URI with a host but no scheme
-                uri = new Uri(publicBaseUrl.Scheme + ":" + uri.AbsolutePath);
-            }
-
-            if (!string.IsNullOrEmpty(uri.Host))
-            {
-                if (!IsSameHost(publicBaseUrl, uri))
-                {
-                    return false;
-                }
-            }
-
-            if (string.IsNullOrEmpty(publicBaseUrl.AbsolutePath) || publicBaseUrl.AbsolutePath == "/")
-            {
-                return true;
-            }
-
-            if (string.IsNullOrEmpty(uri.AbsolutePath) || uri.AbsolutePath == "/")
+            if (!IsSameHost(publicBaseUrl, uriInfo.Uri))
             {
                 return false;
             }
 
-            var basePathText = "/" + publicBaseUrl.GetComponents(UriComponents.Path, UriFormat.UriEscaped).TrimEnd('/');
-            if (!basePathText.EndsWith("/"))
+            // This makes "http://localhost/_dav/" a base path of "/_dav/"
+            var basePath = "/" + publicBaseUrl.GetComponents(UriComponents.Path, UriFormat.UriEscaped);
+            if (!basePath.EndsWith("/"))
             {
-                basePathText += "/";
+                basePath += "/";
             }
 
-            var uriPathText = "/" + uri.GetComponents(UriComponents.Path, UriFormat.UriEscaped).TrimEnd('/');
-            if (!uriPathText.EndsWith("/"))
-            {
-                uriPathText += "/";
-            }
+            var uriPath = uriInfo.PathWithSlash;
 
-            var basePath = new Uri(basePathText);
-            var uriPath = new Uri(uriPathText);
-
-            // This seems to be case-sensitive. Is that really what we want/need?
-            return basePath.IsBaseOf(uriPath);
+            return uriPath.StartsWith(basePath, StringComparison.OrdinalIgnoreCase);
         }
 
-        private Uri? MakeAbsoluteUri(Uri uri, IWebDavContext? context = null)
+        private class UriInfo
         {
-            if (uri.IsAbsoluteUri)
+            public UriInfo(Uri originalUri, Uri? publicBaseUri)
             {
-                return uri;
-            }
+                Scheme = publicBaseUri?.Scheme ?? "http";
 
-            context ??= _contextAccessor?.WebDavContext;
-            if (context == null)
-            {
-                // It is "absolute" in the sense that it is relative to the server root
-                if (uri.OriginalString.StartsWith("/"))
+                var uri = originalUri;
+                if (!uri.IsAbsoluteUri)
                 {
-                    return new Uri(uri.OriginalString);
+                    if (publicBaseUri != null)
+                    {
+                        var basePath = publicBaseUri.OriginalString;
+                        if (!basePath.EndsWith("/"))
+                        {
+                            // Always ensure that the path ends in a slash
+                            publicBaseUri = new Uri(basePath + "/");
+                        }
+
+                        if (uri.OriginalString.StartsWith("///"))
+                        {
+                            uri = new Uri(publicBaseUri, "/" + uri.OriginalString.TrimStart('/'));
+                        }
+                        else if (uri.OriginalString.StartsWith("//"))
+                        {
+                            uri = new Uri(Scheme + "://" + uri.OriginalString.TrimStart('/'));
+                        }
+                        else
+                        {
+                            uri = new Uri(publicBaseUri, uri);
+                        }
+                    }
+                    else
+                    {
+                        if (uri.OriginalString.StartsWith("///"))
+                        {
+                            uri = new Uri("/" + uri.OriginalString.TrimStart('/'));
+                        }
+                        else if (uri.OriginalString.StartsWith("//"))
+                        {
+                            uri = new Uri(Scheme + "://" + uri.OriginalString.TrimStart('/'));
+                        }
+                        else if (uri.OriginalString.StartsWith("/"))
+                        {
+                            uri = new Uri("file://" + uri.OriginalString, UriKind.Absolute);
+                        }
+                    }
                 }
 
-                // Not possible without a context
-                return null;
+                Uri = uri;
+                IsAbsolute = uri.IsAbsoluteUri;
+
+                Path = uri.IsAbsoluteUri
+                    ? "/" + uri.GetComponents(UriComponents.Path, UriFormat.UriEscaped)
+                    : new Uri("file:///" + uri.OriginalString).GetComponents(UriComponents.Path, UriFormat.UriEscaped);
+                PathWithSlash = !Path.EndsWith("/") ? Path + "/" : Path;
             }
 
-            return new Uri(context.PublicBaseUrl, uri);
+            public string Scheme { get; }
+
+            public Uri Uri { get; }
+
+            public bool IsAbsolute { get; }
+
+            public string Path { get; }
+
+            public string PathWithSlash { get; }
         }
     }
 }
