+++ src/FubarDev.WebDavServer/Utils/UriComparisonResult.cs
@@ -15,11 +15,6 @@ namespace FubarDev.WebDavServer.Utils
         PrecedingDifferentHost = -4,
 
         /// <summary>
-        /// The URL points to a different service on the same host
-        /// </summary>
-        PrecedingDifferentService,
-
-        /// <summary>
         /// The URL points to a path that is a preceding sibling of the target path
         /// </summary>
         PrecedingSibling,
@@ -43,11 +38,6 @@ namespace FubarDev.WebDavServer.Utils
         /// The URL points to a path that is a following sibling of the target path
         /// </summary>
         FollowingSibling,
-
-        /// <summary>
-        /// The URL points to a different service on the same host
-        /// </summary>
-        FollowingDifferentService,
 
         /// <summary>
         /// The URL points to different host
